Landing Page
Worked on the landing page based on the work from "Josh Tried Coding". I am liking it but i kind of feel like i need a footer, it looks to freaking plain. I'll see. But i am currently debating on whether to continue with Josh or do the Form builder with Klinton. I am looking to solidify my understanding of server actions so let me do the dashboard with klinton for a form builder. Might add a little AI in the future even.